requirements.txt
openai-whisper
pydub
librosa
soundfile
tqdm
torch
numpy
pandas
